Apr 7, 2026, 1:08 PM

+++ mr_data.jl
@@ -70,7 +70,7 @@ end
 
 pop_tall.age_group = map_pop_group.(pop_tall.age_group)
 pop_tall.population = parse.(Float64,pop_tall.population)
-deaths.rate = parse.(Float,deaths.rate)
+deaths.rate = parse.(Float64,deaths.rate)
 
 
 pop_grouped = DataFrames.combine(groupby(pop_tall,[:tract,:age_group]), :population => sum => :population)
